+++ src/lua/verilua/utils/AgeStaticQueue.lua
@@ -27,9 +27,9 @@ local math_random = math.random
 ---@field pop fun(self: verilua.utils.AgeStaticQueue): T Remove and return the first element from the queue
 ---@field pop_waitable fun(self: verilua.utils.AgeStaticQueue): T Remove and return the first element, waiting if queue is empty
 ---@field wait_not_empty fun(self: verilua.utils.AgeStaticQueue): boolean Wait until the queue is not empty
----@field query_first fun(self: verilua.utils.AgeStaticQueue): T Get the first element without removing it
----@field front fun(self: verilua.utils.AgeStaticQueue): T Alias of query_first
----@field last fun(self: verilua.utils.AgeStaticQueue): T Get the last element without removing it
+---@field query_first fun(self: verilua.utils.AgeStaticQueue): T? Get the first element without removing it, or nil if queue is empty
+---@field front fun(self: verilua.utils.AgeStaticQueue): T? Alias of query_first, or nil if queue is empty
+---@field last fun(self: verilua.utils.AgeStaticQueue): T? Get the last element without removing it, or nil if queue is empty
 ---@field is_empty fun(self: verilua.utils.AgeStaticQueue): boolean Check if the queue is empty
 ---@field is_full fun(self: verilua.utils.AgeStaticQueue): boolean Check if the queue is full
 ---@field size fun(self: verilua.utils.AgeStaticQueue): integer Get current number of elements (alias of used_count)
@@ -170,10 +170,16 @@ function AgeStaticQueue:query_first()
 end
 
 function AgeStaticQueue:front()
+    if self.count == 0 then
+        return nil
+    end
     return self.data[self.first_ptr]
 end
 
 function AgeStaticQueue:last()
+    if self.count == 0 then
+        return nil
+    end
     return self.data[self.last_ptr]
 end